refactor: abstract interface of system contract
